Submit WJLCS-3
Readmd.md is the wiki in Markdown.
Letterset.txt is the letterset file.
Letterset Diagram.xml can be opened using draw.io

Sorry if the wiki goes a little off topic, I kind of felt like I had to get that information out. That and 350 words for what we directly covered in class seemed a little difficult.